Added propper system to handle engine builds
Previously the engine was build into the main project folder, which was an issue regarding module paths and organization.
Now I set up a Build folder, with a debug and publish section. Visual studio is set up to use BuildTask.bat, which copies every cs file, which needs to be included uncompiled with the engine (Module and core source-code) into the engines build folder.