Worked on implementing the first steps towards UI
Added a screenspace shader, that does not do perspective translations, but only recieves a position vector (representing screenspace from (-1.-1) to (1,1) right now.

In the engines update loop I currently use this to render a cube in the top right corner of the screen. This should later be handles by the ui class.

One issue I have not tested yet is depth testing. Everything on ui is currently at z = 0.